+++ lib/DhanHQ/models/forever_order.rb
@@ -67,6 +67,10 @@ module DhanHQ
     #
     class ForeverOrder < BaseModel
       extend DhanHQ::Concerns::BangWrites
+      extend DhanHQ::Concerns::TrackedWrites
+
+      track_class_writes :create
+      track_writes :modify, :cancel
 
       bang_class_writes :create
       bang_writes :modify, :cancel