+++ lib/DhanHQ/models/global_stocks/order.rb
@@ -37,6 +37,10 @@ module DhanHQ
       #
       class Order < BaseModel
         extend DhanHQ::Concerns::BangWrites
+        extend DhanHQ::Concerns::TrackedWrites
+
+        track_class_writes :place
+        track_writes :modify, :cancel
 
         bang_class_writes :place
         bang_writes :modify, :cancel