+++ lib/DhanHQ/models/iceberg_order.rb
@@ -49,6 +49,10 @@ module DhanHQ
     #
     class IcebergOrder < BaseModel
       extend DhanHQ::Concerns::BangWrites
+      extend DhanHQ::Concerns::TrackedWrites
+
+      track_class_writes :create
+      track_writes :modify, :cancel
 
       bang_class_writes :create
       bang_writes :modify, :cancel